fix: Correct constraints and tune cost function
The signs of some angular values needed to be changed for the controller to work properly. Positive angles are measured counter-clockwise in the vehicle model, but negative steering values actuate counter-clockwise steering of the vehicle. The controller now works exclusively with positive angles being measured counter-clockwise internally, and the sign of the steering angle output by the optimizer is being inverted in order to send the correct steering command to the simulator.

The cost function was tuned to compute smoother steering commands.

Also, main.cpp now compensates latency by predicting the initial state after latency. This leads to much smoother driving behavior.